refactor: 3단계 지하철구간 관리
- 도메인 호출 api 패키지 위치 이동
